make loop starts/ends draggable
