+++ mamar-web/src/app/doc/SegmentMap.tsx
@@ -1,40 +1,12 @@
 import classNames from "classnames"
-import { Segment } from "pm64-typegen"
 import { useId } from "react"
 
+import Ruler, { ticksToStyle, useSegmentLengths } from "./Ruler"
 import styles from "./SegmentMap.module.scss"
 
 import TrackControls from "../emu/TrackControls"
 import { useBgm, useDoc, useVariation } from "../store"
 import useSelection, { SelectionProvider } from "../util/hooks/useSelection"
-
-interface Loop {
-    start: number
-    end: number
-}
-
-function getLoops(segments: Segment[]): Loop[] {
-    const loops = []
-
-    for (let startIdx = 0; startIdx < segments.length; startIdx++) {
-        const start = segments[startIdx]
-        if (start.type === "StartLoop") {
-            // Look for EndLoop
-            for (let endIdx = 0; endIdx < segments.length; endIdx++) {
-                const end = segments[endIdx]
-                if (end.type === "EndLoop" && end.label_index === start.label_index) {
-                    loops.push({
-                        start: startIdx,
-                        end: endIdx,
-                    })
-                    break
-                }
-            }
-        }
-    }
-
-    return loops
-}
 
 function PianoRollThumbnail({ trackIndex, trackListIndex }: { trackIndex: number, trackListIndex: number }) {
     const [doc, dispatch] = useDoc()
@@ -81,96 +53,12 @@ function PianoRollThumbnail({ trackIndex, trackListIndex }: { trackIndex: number
     }
 }
 
-function ticksToStyle(ticks: number) {
-    // TODO: zoom via a css variable
-    return {
-        width: `calc(${ticks}px / var(--ruler-zoom))`,
-    }
-}
-
-function LoopHandle() {
-    // TODO
-    return <div>
-    </div>
-}
-
-// TODO: bar counts where a segment is not a full bar
-function Ruler({ segmentLengths, loops }: { segmentLengths: number[], loops: Loop[] }) {
-    const TICKS_PER_BEAT = 48
-    const BEATS_PER_BAR = 4 // TODO: read time signature from midi
-
-    const segments = []
-    let time = 0
-    let inLoop = false
-    for (let segment = 0; segment < segmentLengths.length; segment++) {
-        const length = segmentLengths[segment]
-
-        if (length === 0) {
-            // Loop or other, so check for loop handle
-            for (const { start, end } of loops) {
-                if (segment === start) {
-                    inLoop = true
-                    segments.push(<LoopHandle />)
-                }
-                if (segment === end) {
-                    inLoop = false
-                    segments.push(<LoopHandle />)
-                }
-                continue
-            }
-            continue
-        }
-
-        const barLength = length / (TICKS_PER_BEAT * BEATS_PER_BAR)
-        const bars = []
-        for (let i = 0; i < barLength; i++) {
-            const bar = Math.floor(time / (TICKS_PER_BEAT * BEATS_PER_BAR)) + i
-            bars.push(<div className={styles.bar} style={ticksToStyle(TICKS_PER_BEAT * BEATS_PER_BAR)}>
-                {bar}
-            </div>)
-        }
-
-        segments.push(<div
-            className={classNames({
-                [styles.rulerSegment]: true,
-                [styles.loop]: inLoop,
-            })}
-            style={ticksToStyle(length)}
-        >
-            {bars}
-        </div>)
-
-        time += length
-    }
-
-    return <div className={styles.ruler}>
-        {segments}
-    </div>
-}
-
 function Container() {
     const [variation] = useVariation()
     const selection = useSelection()
+    const segmentLengths = useSegmentLengths()
 
-    const tracks = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15]
-    const segments = variation?.segments ?? []
-    const loops = getLoops(segments)
-
-    const [bgm] = useBgm()
-    const segmentLengths = segments.map(segment => {
-        if (bgm && segment.type === "Subseg") {
-            const master = bgm.trackLists[segment.trackList].tracks[0]
-            return master.commands.vec.reduce((totalDelay, event) => {
-                if (event.type === "Delay") {
-                    return totalDelay + event.value
-                } else {
-                    return totalDelay
-                }
-            }, 0)
-        } else {
-            return 0
-        }
-    })
+    const tracks = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15] // note: no master track
 
     return <div
         aria-label="Segments"
@@ -180,10 +68,10 @@ function Container() {
         }}
     >
         {variation && <div>
-            <Ruler segmentLengths={segmentLengths} loops={loops} />
-            {tracks.map(i => <tr key={i} className={styles.track} aria-label={`Track ${i+1}`}>
+            <Ruler segments={variation.segments} />
+            {tracks.map(i => <div key={i} className={styles.track} aria-label={`Track ${i}`}>
                 {<div className={styles.trackHead}>
-                    <div className={styles.trackName}>Track {i + 1}</div>
+                    <div className={styles.trackName}>Track {i}</div>
                     <TrackControls trackIndex={i} />
                 </div>}
                 {variation.segments.map((segment, segmentIndex) => {
@@ -198,7 +86,7 @@ function Container() {
                         return <div key={segment.id} />
                     }
                 })}
-            </tr>)}
+            </div>)}
         </div>}
     </div>
